chore: migrate setupFile to globalSetup
