+++ packages/client/vitest.config.mts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    setupFiles: 'vitest.setup.ts',
+    globalSetup: 'globalSetup.ts',
     coverage: {
       provider: 'istanbul',
       reporter: ['text', 'json-summary', 'json', 'html'],
